+++ client/flutter/divkit/lib/src/core/widgets/image/div_image_widget.dart
@@ -49,6 +49,7 @@ class _DivImageWidgetState extends State<DivImageWidget> {
         data: widget.data,
         action: widget.data.action,
         actions: widget.data.actions,
+        longtapActions: widget.data.longtapActions,
         child: StreamBuilder<DivImageModel>(
           stream: stream,
           builder: (context, snapshot) {